requirements.txt
python-telegram-bot
# نصب نسخه Master برای حل مشکل n-challenge
yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git@master